% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{integrated_cd_auc}
\alias{integrated_cd_auc}
\title{Calculate integrated C/D AUC}
\usage{
integrated_cd_auc(
  y_true = NULL,
  risk = NULL,
  surv = NULL,
  times = NULL,
  auc = NULL
)
}
\arguments{
\item{y_true}{a \code{survival::Surv} object containing the times and statuses of observations for which the metric will be evaluated}

\item{risk}{ignored, left for compatibility with other metrics}

\item{surv}{a matrix containing the predicted survival functions for the considered observations, each row represents a single observation, whereas each column one time point}

\item{times}{a vector of time points at which the survival function was evaluated}

\item{auc}{a vector containing already calculated AUC metric at the time points specified in the times parameter. If this is provided all arguments except \code{times} and \code{auc} are ignored}
}
\value{
numeric from 0 to 1, higher values indicate better performance

#' @section References:
\itemize{
\item [1] Uno, Hajime, et al. \href{https://www.jstor.org/stable/27639883#metadata_info_tab_contents}{"Evaluating prediction rules for t-year survivors with censored regression models."} Journal of the American Statistical Association 102.478 (2007): 527-537.
\item [2] Hung, Hung, and Chin‐Tsang Chiang. \href{https://www.jstor.org/stable/41000414#metadata_info_tab_contents}{"Optimal composite markers for time‐dependent receiver operating characteristic curves with censored survival data."} Scandinavian Journal of Statistics 37.4 (2010): 664-679.
}
}
\description{
This function calculates the integrated Cumulative/Dynamic AUC metric for a survival model.
}
\details{
It is useful to see how a model performs as a whole, not at specific time points, for example for easier comparison. This function allows for calculating the integral of the C/D AUC metric numerically using the trapezoid method.
}
\examples{

library(survival)
library(survex)

cph <- coxph(Surv(time, status) ~ ., data = veteran, model = TRUE, x = TRUE, y = TRUE)
cph_exp <- explain(cph)

y <- cph_exp$y
times <- cph_exp$times
surv <- cph_exp$predict_survival_function(cph, cph_exp$data, times)


# calculating directly
integrated_cd_auc(y, surv = surv, times = times)

# calculating based on given auc vector
auc <- cd_auc(y, surv = surv, times = times)
integrated_cd_auc(times = times, auc = auc)

}
\seealso{
\code{\link[=cd_auc]{cd_auc()}} \code{\link[=loss_one_minus_cd_auc]{loss_one_minus_cd_auc()}}
}
