% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{loss_one_minus_integrated_cd_auc}
\alias{loss_one_minus_integrated_cd_auc}
\title{Calculate integrated C/D AUC loss}
\usage{
loss_one_minus_integrated_cd_auc(
  y_true = NULL,
  risk = NULL,
  surv = NULL,
  times = NULL,
  auc = NULL
)
}
\arguments{
\item{y_true}{a \code{survival::Surv} object containing the times and statuses of observations for which the metric will be evaluated}

\item{risk}{ignored, left for compatibility with other metrics}

\item{surv}{a matrix containing the predicted survival functions for the considered observations, each row represents a single observation, whereas each column one time point}

\item{times}{a vector of time points at which the survival function was evaluated}

\item{auc}{a vector containing already calculated AUC metric at the time points specified in the times parameter. If this is provided all arguments except \code{times} and \code{auc} are ignored}
}
\value{
numeric from 0 to 1, lower values indicate better performance

#' @section References:
\itemize{
\item \href{https://www.jstor.org/stable/27639883#metadata_info_tab_contents}{[1] Uno, Hajime, et al. "Evaluating prediction rules for t-year survivors with censored regression models." Journal of the American Statistical Association 102.478 (2007): 527-537.}
\item \href{https://www.jstor.org/stable/41000414#metadata_info_tab_contents}{[2] Hung, Hung, and Chin‐tsang Chiang. "Optimal composite markers for time‐dependent receiver operating characteristic curves with censored survival data." Scandinavian journal of statistics 37.4 (2010): 664-679.}
\item \href{https://scikit-survival.readthedocs.io/en/stable/api/generated/sksurv.metrics.cumulative_dynamic_auc.html}{[3] scikit-survival documentation entry}
}
}
\description{
This function subtracts integrated the C/D AUC metric from one to obtain a loss function whose lower values indicate better model performance (useful for permutational feature importance)
}
\examples{

library(survival)
library(survex)

cph <- coxph(Surv(time, status) ~ ., data = veteran, model = TRUE, x = TRUE, y = TRUE)
cph_exp <- explain(cph)

y <- cph_exp$y
times <- cph_exp$times
surv <- cph_exp$predict_survival_function(cph, cph_exp$data, times)


# calculating directly
loss_one_minus_integrated_cd_auc(y, surv = surv, times = times)

# calculating based on given auc vector
auc <- cd_auc(y, surv = surv, times = times)
loss_one_minus_integrated_cd_auc(times = times, auc = auc)

}
\seealso{
\code{\link[=integrated_cd_auc]{integrated_cd_auc()}} \code{\link[=cd_auc]{cd_auc()}} \code{\link[=loss_one_minus_cd_auc]{loss_one_minus_cd_auc()}}
}
