% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{model_diagnostics}
\alias{model_diagnostics}
\alias{model_diagnostics.surv_explainer}
\title{Dataset Level Model Diagnostics}
\usage{
model_diagnostics(explainer, ...)

\method{model_diagnostics}{surv_explainer}(explainer, ..., output_type = "chf")
}
\arguments{
\item{explainer}{an explainer object - model preprocessed by the \code{explain()} function}

\item{...}{other parameters passed to \code{DALEX::model_diagnostics} if \code{output_type == "risk"}, otherwise passed to internal functions.}

\item{output_type}{either \code{"chf"}, \code{"survival"} or \code{"risk"} the type of survival model output that should be used for explanations. If \code{"chf"} or \code{"survival"} the explanations are based on survival residuals. Otherwise the scalar risk predictions are used by the \code{DALEX::model_diagnostics} function.}
}
\value{
An object of class \code{c("model_diagnostics_survival")}. It's a list with the explanations in the \code{result} element.
}
\description{
This function calculates martingale and deviance residuals.
}
\examples{
\donttest{
library(survival)
library(survex)

cph <- coxph(Surv(time, status) ~ ., data = veteran, model = TRUE, x = TRUE, y = TRUE)
rsf_ranger <- ranger::ranger(Surv(time, status) ~ .,
  data = veteran,
  respect.unordered.factors = TRUE,
  num.trees = 100,
  mtry = 3,
  max.depth = 5
)

cph_exp <- explain(cph)

rsf_ranger_exp <- explain(rsf_ranger,
  data = veteran[, -c(3, 4)],
  y = Surv(veteran$time, veteran$status)
)

TODO kod

}
}
