% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_survshap.R
\name{model_survshap}
\alias{model_survshap}
\alias{model_survshap.surv_explainer}
\title{Global SHAP Values}
\usage{
model_survshap(explainer, ...)

\method{model_survshap}{surv_explainer}(
  explainer,
  calculation_method = "kernelshap",
  aggregation_method = "integral",
  new_observation = NULL,
  y_true = NULL,
  ...,
  N = NULL
)
}
\arguments{
\item{explainer}{an explainer object - model preprocessed by the \code{explain()} function}

\item{...}{additional parameters, passed to internal functions}

\item{calculation_method}{a character, either \code{"kernelshap"} for use of \code{kernelshap} library (providing faster Kernel SHAP with refinements), \code{"exact_kernel"} for exact Kernel SHAP estimation,
or \code{"treeshap"} for use of \code{treeshap} library (efficient implementation to compute SHAP values for tree-based models).}

\item{aggregation_method}{a character, either \code{"mean_absolute"} or \code{"integral"}, \code{"max_absolute"}, \code{"sum_of_squares"}}

\item{new_observation}{new observations for which predictions need to be explained}

\item{y_true}{a two element numeric vector or matrix of one row and two columns, the first element being the true observed time and the second the status of the observation, used for plotting}

\item{N}{A positive integer indicating the number of observations that should be used to compute global SHAP values.}
}
\value{
An object of class \code{aggregated_surv_shap} containing the computed global SHAP values.
}
\description{
This function computes global SHAP values.
}
\details{
If specifying \code{y_true}, also \code{new_observation} must be specified.
Using the argument \code{new_observation}, global SHAP values are computed for the provided data. Otherwise,
global SHAP values are computed for the data, the \code{explainer} was trained with.
}
