% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predict_profile_survival.R
\name{plot.predict_profile_survival}
\alias{plot.predict_profile_survival}
\title{Plot Predict Profile for Survival Models}
\usage{
\method{plot}{predict_profile_survival}(x, ...)
}
\arguments{
\item{x}{an object of class  \code{"predict_profile_survival"} to be plotted}

\item{...}{additional parameters passed to the \code{plot.surv_ceteris_paribus} function}
}
\value{
A grid of \code{ggplot} objects arranged with the \code{gridExtra::grid.arrange} function.
}
\description{
This function plots objects of class \code{"predict_profile_survival"} - local explanations
for survival models created using the \code{predict_profile()} function.
}
\section{Plot options}{

\subsection{\code{plot.surv_ceteris_paribus}}{
\itemize{
\item \code{x} - an object of class \code{predict_profile_survival} to be plotted
\item \code{...} - additional parameters, unused, currently ignored
\item \code{colors} - character vector containing the colors to be used for plotting variables (containing either hex codes "#FF69B4", or names "blue")
\item \code{variable_type} - character, either \code{"numerical"}, \code{"categorical"} or \code{NULL} (default), select only one type of variable for plotting, or leave \code{NULL} for all
\item \code{facet_ncol} - number of columns for arranging subplots
\item \code{variables} - character, names of the variables to be plotted
\item \code{numerical_plot_type} - character, either \code{"lines"}, or \code{"contours"} selects the type of numerical variable plots
\item \code{title} - character, title of the plot
\item \code{subtitle} - character, subtitle of the plot, \code{'default'} automatically generates "created for XXX, YYY models", where XXX and YYY are the explainer labels
\item \code{rug} - character vector. One of "all", "events", "censors", "none" or NULL. Which times to mark on the x axis in \code{geom_rug()}.
\item \code{rug_colors} - character vector containing two colors (containing either hex codes "#FF69B4", or names "blue"). The first color (red by default) will be used to mark event times, whereas the second (grey by default) will be used to mark censor times.
}
}
}

\examples{
\donttest{
library(survival)
library(survex)

model <- randomForestSRC::rfsrc(Surv(time, status) ~ ., data = veteran)
exp <- explain(model)

p_profile <- predict_profile(exp, veteran[1, -c(3, 4)])

plot(p_profile)

p_profile_with_cat <- predict_profile(
    exp,
    veteran[1, -c(3, 4)],
    categorical_variables = c("trt", "prior")
)

plot(p_profile_with_cat)
}
}
\seealso{
Other functions for plotting 'predict_profile_survival' objects: 
\code{\link{plot.surv_ceteris_paribus}()}
}
\concept{functions for plotting 'predict_profile_survival' objects}
