% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_model_profiles.R
\name{surv_aggregate_profiles}
\alias{surv_aggregate_profiles}
\title{Helper functions for \code{model_profile.R}}
\usage{
surv_aggregate_profiles(
  x,
  ...,
  variable_type = NULL,
  groups = NULL,
  variables = NULL,
  center = FALSE
)
}
\arguments{
\item{x}{an object containing calculated ceteris_paribus profiles}

\item{...}{other parameters, ignored}

\item{variable_type}{character, either \code{"numerical"} or \code{"categorical"}, the type of variable to be calculated, if left \code{NULL} (default), both are calculated}

\item{variables}{a character vector containing names of variables to be explained}

\item{center}{logical, if the profiles should be centered before aggregations}
}
\value{
A data.frame with calculated results.
}
\description{
Helper functions for \code{model_profile.R}
}
\keyword{internal}
