% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ceteris_paribus.R
\name{surv_ceteris_paribus}
\alias{surv_ceteris_paribus}
\alias{surv_ceteris_paribus.surv_explainer}
\title{Helper functions for \code{predict_profile.R}}
\usage{
surv_ceteris_paribus(x, ...)

\method{surv_ceteris_paribus}{surv_explainer}(
  x,
  new_observation,
  variables = NULL,
  categorical_variables = NULL,
  variable_splits = NULL,
  grid_points = 101,
  variable_splits_type = "uniform",
  center = FALSE,
  output_type = "survival",
  ...
)
}
\arguments{
\item{x}{an explainer object - model preprocessed by the \code{explain()} function}

\item{...}{other parameters, currently ignored}

\item{new_observation}{a new observation for which predictions need to be explained}

\item{variables}{character, names of the variables to be included in the calculations}

\item{categorical_variables}{character vector, names of variables that should be treated as categories (factors are included by default)}

\item{variable_splits}{named list of splits for variables, in most cases created with internal functions. If NULL then it will be calculated based on validation data available in the explainer}

\item{grid_points}{maximum number of points for profile calculations. Note that the final number of points may be lower than grid_points. Will be passed to internal function. By default \code{101}.}

\item{variable_splits_type}{character, decides how variable grids should be calculated. Use \code{"quantiles"} for percentiles or \code{"uniform"} (default) to get uniform grid of points.}
}
\value{
A data.frame containing the result of the calculation.
}
\description{
Helper functions for \code{predict_profile.R}

Helper functions for \code{predict_profile.R}
}
\keyword{internal}
