% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_shap.R
\name{surv_shap}
\alias{surv_shap}
\title{Helper functions for \code{predict_parts.R}}
\usage{
surv_shap(
  explainer,
  new_observation,
  ...,
  y_true = NULL,
  calculation_method = "kernelshap",
  aggregation_method = "integral",
  observation_aggregation_method = mean,
  path = "average",
  B = 25,
  exact = FALSE
)
}
\arguments{
\item{explainer}{an explainer object - model preprocessed by the \code{explain()} function}

\item{new_observation}{new observations for which predictions need to be explained}

\item{...}{additional parameters, passed to internal functions}

\item{y_true}{a two element numeric vector or matrix of one row and two columns, the first element being the true observed time and the second the status of the observation, used for plotting}

\item{calculation_method}{a character, either \code{"kernelshap"} for use of \code{kernelshap} library (providing faster Kernel SHAP with refinements) or \code{"exact_kernel"} for exact Kernel SHAP estimation}

\item{aggregation_method}{a character, either \code{"mean_absolute"} or \code{"integral"}, \code{"max_absolute"}, \code{"sum_of_squares"}}

\item{observation_aggregation_method}{a function, if \code{new_observation} contains multiple observation this function is applied to the same time point of generated shap profiles for each observation. Defaults to \code{mean}.}
}
\value{
A list, containing the calculated SurvSHAP(t) results in the \code{result} field
}
\description{
Helper functions for \code{predict_parts.R}
}
\section{References}{

\itemize{
\item [1] Krzyziński, Mateusz, et al. \href{https://www.sciencedirect.com/science/article/pii/S0950705122013302}{"SurvSHAP(t): Time-dependent explanations of machine learning survival models."} Knowledge-Based Systems 262 (2023): 110234
}
}

\keyword{internal}
