% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_survshap.R
\name{model_survshap}
\alias{model_survshap}
\alias{model_survshap.surv_explainer}
\title{Global SHAP Values}
\usage{
model_survshap(explainer, ...)

\method{model_survshap}{surv_explainer}(
  explainer,
  new_observation = NULL,
  y_true = NULL,
  calculation_method = "kernelshap",
  aggregation_method = "integral",
  ...
)
}
\arguments{
\item{explainer}{an explainer object - model preprocessed by the \code{explain()} function}

\item{...}{additional parameters, passed to internal functions}

\item{new_observation}{new observations for which predictions need to be explained}

\item{y_true}{a two element numeric vector or matrix of one row and two columns, the first element being the true observed time and the second the status of the observation, used for plotting}

\item{calculation_method}{a character, either \code{"kernelshap"} for use of \code{kernelshap} library (providing faster Kernel SHAP with refinements) or \code{"exact_kernel"} for exact Kernel SHAP estimation}

\item{aggregation_method}{a character, either \code{"integral"}, \code{"integral_absolute"}, \code{"mean_absolute"}, \code{"max_absolute"}, or \code{"sum_of_squares"}}
}
\value{
An object of class \code{aggregated_surv_shap} containing the computed global SHAP values.
}
\description{
This function computes global SHAP values.
}
\details{
If specifying \code{y_true}, also \code{new_observation} must be specified.
Using the argument \code{new_observation}, global SHAP values are computed for the provided data. Otherwise,
global SHAP values are computed for the data, the \code{explainer} was trained with.
}
\examples{
\donttest{
veteran <- survival::veteran
rsf_ranger <- ranger::ranger(
  survival::Surv(time, status) ~ .,
  data = veteran,
  respect.unordered.factors = TRUE,
  num.trees = 100,
  mtry = 3,
  max.depth = 5
)
rsf_ranger_exp <- explain(
  rsf_ranger,
  data = veteran[, -c(3, 4)],
  y = Surv(veteran$time, veteran$status),
  verbose = FALSE
)

ranger_global_survshap <- model_survshap(
  explainer = rsf_ranger_exp,
  new_observation = veteran[1:40, !colnames(veteran) \%in\% c("time", "status")],
  y_true = Surv(veteran$time[1:40], veteran$status[1:40]),
  aggregation_method = "mean_absolute",
  calculation_method = "kernelshap",
  N = 5L
)
plot(ranger_global_survshap)
}

}
