% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surv_shap.R
\name{plot.aggregated_surv_shap}
\alias{plot.aggregated_surv_shap}
\title{Plot Aggregated SurvSHAP(t) Explanations for Survival Models}
\usage{
\method{plot}{aggregated_surv_shap}(
  x,
  kind = "importance",
  ...,
  title = "default",
  subtitle = "default",
  max_vars = 7,
  colors = NULL
)
}
\arguments{
\item{x}{an object of class \code{aggregated_surv_shap} to be plotted}

\item{kind}{character, one of \code{"importance"}, \code{"swarm"}, or \code{"profile"}. Type of chart to be plotted; \code{"importance"} shows the importance of variables over time and aggregated, \code{"swarm"} shows the distribution of SurvSHAP(t) values for variables and observations, \code{"profile"} shows the dependence of SurvSHAP(t) values on variable values.}

\item{...}{additional parameters passed to internal functions}

\item{title}{character, title of the plot}

\item{subtitle}{character, subtitle of the plot, \code{'default'} automatically generates "created for the XXX model (n = YYY)", where XXX is the explainer label and YYY is the number of observations used for calculations}

\item{max_vars}{maximum number of variables to be plotted (least important variables are ignored), by default 7}

\item{colors}{character vector containing the colors to be used for plotting variables (containing either hex codes "#FF69B4", or names "blue")}
}
\value{
An object of the class \code{ggplot}.
}
\description{
This functions plots objects of class \code{aggregated_surv_shap} - aggregated time-dependent
explanations of survival models created using the \code{model_survshap()} function.
}
\section{Plot options}{

\subsection{\code{plot.aggregated_surv_shap(type = "importance")}}{
\itemize{
\item \code{rug} - character, one of \code{"all"}, \code{"events"}, \code{"censors"}, \code{"none"} or \code{NULL}. Which times to mark on the x axis in \code{geom_rug()}.
\item \code{rug_colors} - character vector containing two colors (containing either hex codes "#FF69B4", or names "blue"). The first color (red by default) will be used to mark event times, whereas the second (grey by default) will be used to mark censor times.
\item \verb{xlab_left, ylab_right} - axis labels for left and right plots (due to different aggregation possibilities)
}
}

\subsection{\code{plot.aggregated_surv_shap(type = "swarm")}}{
\itemize{
\item no additional parameters
}
}

\subsection{\code{plot.aggregated_surv_shap(type = "swarm")}}{
\itemize{
\item \code{variable} - variable for which the profile is to be plotted, by default first from result data
\item \code{color_variable} - variable used to denote the color, by default equal to \code{variable}
}
}
}

\examples{
\donttest{
library(survival)
library(survex)

model <- randomForestSRC::rfsrc(Surv(time, status) ~ ., data = veteran)
exp <- explain(model)

p_parts_shap <- predict_parts(exp, veteran[1, -c(3, 4)], type = "survshap")
plot(p_parts_shap)
}

}
