% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_diagnostics_survival.R
\name{plot.model_diagnostics_survival}
\alias{plot.model_diagnostics_survival}
\title{Plot Model Diagnostics for Survival Models}
\usage{
\method{plot}{model_diagnostics_survival}(
  x,
  ...,
  type = "deviance",
  xvariable = "index",
  smooth = as.logical(xvariable != "index"),
  title = "Model diagnostics",
  subtitle = "default",
  facet_ncol = NULL,
  colors = c("#160e3b", "#f05a71", "#ceced9")
)
}
\description{
Plot Model Diagnostics for Survival Models
}
