% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_profile_survival.R
\name{plot2.model_profile_survival}
\alias{plot2.model_profile_survival}
\title{Plot Model Profile for Survival Models (without continuous time aspect)}
\usage{
\method{plot2}{model_profile_survival}(
  x,
  variable,
  times = NULL,
  marginalize_over_time = FALSE,
  plot_type = "pdp+ice",
  ...,
  title = "Partial dependence profile",
  subtitle = "default",
  colors = NULL
)
}
\arguments{
\item{x}{an object of class \code{model_profile_survival} to be plotted}

\item{variable}{character, name of a single variable to be plotted}

\item{times}{numeric vector, times for which the profile should be plotted, the times must be present in the 'times' field of the explainer. If \code{NULL} (default) then the median time from the explainer object is used.}

\item{marginalize_over_time}{logical, if \code{TRUE} then the profile is calculated for all times and then averaged over time, if \code{FALSE} (default) then the profile is calculated for each time separately}

\item{plot_type}{character, one of \code{"pdp"}, \code{"ice"}, \code{"pdp+ice"} selects the type of plot to be drawn}

\item{...}{other parameters. Currently ignored.}

\item{title}{character, title of the plot}

\item{subtitle}{character, subtitle of the plot, \code{'default'} automatically generates "created for XXX, YYY models", where XXX and YYY are the explainer labels}

\item{colors}{character vector containing the colors to be used for plotting variables (containing either hex codes "#FF69B4", or names "blue")}
}
\value{
A \code{ggplot} object.
}
\description{
This function plots objects of class \code{"model_profile_survival"} created
using the \code{model_profile()} function.
}
\examples{
\donttest{
library(survival)
library(survex)

model <- randomForestSRC::rfsrc(Surv(time, status) ~ ., data = veteran)
exp <- explain(model)

m_prof <- model_profile(exp, categorical_variables = "trt")

plot2(m_prof_cph, variable = "karno", plot_type = "pdp+ice")

plot2(m_prof_cph, times = c(1, 2.72), variable = "karno", plot_type = "pdp+ice")

plot2(m_prof_cph, times = c(1, 2.72), variable = "celltype", plot_type = "pdp+ice")
}

}
