% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_model_performance.R
\name{surv_model_performance}
\alias{surv_model_performance}
\title{Helper functions for \code{model_performance.R}}
\usage{
surv_model_performance(explainer, ..., times = NULL, type = "metrics")
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the \code{explain} function}

\item{...}{other parameters, currently ignored}

\item{times}{a numeric vector, time points at which ROC curves are calculated if \code{type == "roc"} or at which metrics are calculated if \code{type == "metrics"}. Note: if \code{type=="roc"} this parameter is obligatory}

\item{type}{character, either \code{"metrics"} which calculates performance metrics or \code{"roc"} which calculates ROC curves at specific time points}
}
\value{
Either a list when \code{type == "metrics"} or a data.frame if \code{type == "roc"}
}
\description{
Helper functions for \code{model_performance.R}
}
\keyword{internal}
